feat: 로그인 여부 검증 구현
