+++ src/main/java/webserver/controller/UserController.java
@@ -7,12 +7,14 @@ import model.request.HttpRequest;
 import model.annotation.Api;
 import model.response.HttpResponse;
 import utils.builder.ResponseBuilder;
+import utils.utils.LoginUtils;
 import webserver.dao.UserDao;
 import webserver.infra.ViewResolver;
 import webserver.service.UserService;
 
 import static constant.DefaultConstant.*;
 import static constant.HeaderConstant.*;
+import static utils.utils.LoginUtils.*;
 import static utils.utils.TemplateUtils.*;
 
 public class UserController extends ApiController {
@@ -40,14 +42,10 @@ public class UserController extends ApiController {
 
     @Api(method = HttpMethod.GET, url = "/user/list.html")
     public HttpResponse showUserList(HttpRequest request){
-        if (cookieNotExists(request)) {
+        if (!isLogin(request)) {
             return ResponseBuilder.found(DEFAULT_PAGE);
         }
 
         return ViewResolver.resolve(handleUserListTemplate());
-    }
-
-    private boolean cookieNotExists(HttpRequest request) {
-        return request.findHeaderValue(COOKIE, null) == null;
     }
 }
